deps: bump aikit v1.4.0 → v1.5.0; make int8 the reranker default
aikit v1.5.0 fixes the q8 reranker (pooled scratch + dequant-once-then-SIMD
matmul — the pre-fix path re-widened int8→f32 inside the GEMM and allocated
4.4 GiB of scratch, making it ~5× slower than f32). On ken's rerank path
int8 now reaches f32 latency PARITY (50-doc cold: 7.35s vs 7.75s, arm64) at
~21× less runtime memory (18 MiB vs 379 MiB) and ¼ the weight storage
(~140 MB resident vs ~547 MB), cosine 0.997 vs f32 unchanged.

So int8 is strictly better-or-equal → it's now the default:
- KEN_MCP_RERANK_QUANT default f32 → int8 (cmd/ken-mcp setupReranker).
- --rerank-quant default f32 → int8 (cmd/ken attachRerankerIfNeeded);
  unknown values still error, the cache filename still scopes per-quant.
No re-download: LoadQ8 quantizes the existing CodeRankEmbed snapshot
in-process, so the same ~/.ken/rerank-model works.

Reverses the v1.0.1-era "int8 slower on Apple Silicon" notes in
DEVELOPERS.md + PERF-expectations.md (that was the pre-fix q8 path).
Validated against the published v1.5.0 tag (GOWORK=off): full suite green;
new TestSetupReranker_DefaultQuantIsInt8 locks the default in.

+++ cmd/ken-mcp/main.go
@@ -728,7 +728,9 @@ func setupReranker(logger *kenmcp.Logger) (*search.LazyReranker, *rerankerLoader
 	topN := envInt("KEN_MCP_RERANK_TOP_N", 50, logger)
 	cacheSize := envInt("KEN_MCP_RERANK_CACHE_SIZE", search.DefaultRerankerCacheSize, logger)
 	beta := envFloat("KEN_MCP_RERANK_BETA", 0.25, logger)
-	quant := envEnum("KEN_MCP_RERANK_QUANT", []string{"f32", "int8"}, "f32", logger)
+	// Default int8: aikit ≥v1.5.0's q8 reranker reaches f32 latency parity at
+	// ~21× less runtime memory + ¼ weight storage, cosine 0.997 unchanged.
+	quant := envEnum("KEN_MCP_RERANK_QUANT", []string{"f32", "int8"}, "int8", logger)
 	adaptiveThreshold, adaptiveMinN := parseRerankAdaptive(logger)
 
 	if modelDir == "" {
